requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.4.0
django-allauth==0.61.1
dj-rest-auth==5.0.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
pillow==10.2.0
channels==4.0.0
daphne==4.0.0